refactor(nim-serving): organize src into components, hooks, and k8s folders
Move React components to components/, custom hooks to hooks/, and K8s
utilities to k8s/ for better package structure as the module grows.
