+++ packages/nim-serving/src/k8s/nimK8sUtils.ts
@@ -7,14 +7,14 @@ import { NIMAccountKind, SecretKind, K8sCondition } from '@odh-dashboard/interna
 import { SecretModel } from '@odh-dashboard/internal/api/models';
 import { createSecret, getSecret, replaceSecret } from '@odh-dashboard/internal/api/k8s/secrets';
 import { allSettledPromises } from '@odh-dashboard/internal/utilities/allSettledPromises';
-import { NIMAccountModel, listNIMAccounts } from './k8s/nimAccounts';
+import { NIMAccountModel, listNIMAccounts } from './nimAccounts';
 import {
   NIM_SECRET_GENERATE_NAME,
   NIM_ACCOUNT_NAME,
   NIM_API_KEY_DATA_KEY,
   NGC_API_KEY_DATA_KEY,
   NIM_FORCE_VALIDATION_ANNOTATION,
-} from './nimConstants';
+} from '../nimConstants';
 
 export const assembleNIMSecret = (namespace: string, apiKey: string): SecretKind => ({
   apiVersion: 'v1',
